+++ CHANGELOG.md
@@ -7,11 +7,14 @@
 - Hide text field and "send" button
 - Swap out cURL hints for nak ones on empty pages
 - Don't fetch older messages on new subscriptions
+- Changed kind from 30078 to 7741
+- Don't let encrypted notifications bypass whitelist
 
 ### Added
 
-- author npub on notifications
-- public/private indicator
+- Author npub on notifications
+- Public/private indicator
+- Amber integration for addressed events
 
 ## v0.1.0